requirements.txt
fastapi
uvicorn
pymongo
praw
openai
apscheduler
python-dotenv
mongoengine
azure-storage-queue